requirements.txt
fastapi
uvicorn
openai
python-dotenv
langgraph
langchain-openai
langchain
langchain-core
langchain-mcp-adapters
pydantic
sqlalchemy
chromadb
langchain-community
